refactoring[st]: Reorganizies feedback and submission service.
