--- SmartTutor/ContentModel/IChallengeService.cs
@@ -1,9 +1,0 @@
-﻿using SmartTutor.ContentModel.LearningObjects.Challenges;
-
-namespace SmartTutor.ContentModel
-{
-    public interface IChallengeService
-    {
-        ChallengeEvaluation EvaluateSubmission(string[] sourceCode, int challengeId, string traineeId);
-    }
-}
