+++ SmartTutor/ProgressModel/Feedback/Repository/FeedbackDatabaseRepository.cs
@@ -1,6 +1,6 @@
-using System.Linq;
 using Microsoft.EntityFrameworkCore;
 using SmartTutor.Database;
+using System.Linq;
 
 namespace SmartTutor.ProgressModel.Feedback.Repository
 {
@@ -13,19 +13,16 @@ namespace SmartTutor.ProgressModel.Feedback.Repository
             _dbContext = dbContext;
         }
 
-        public void SaveFeedback(LearningObjectFeedback feedback)
+        public void SaveOrUpdate(LearningObjectFeedback feedback)
         {
-            _dbContext.LearningObjectFeedbacks.Add(feedback);
+            var loadedFeedback = Get(feedback.LearningObjectId, feedback.LearnerId);
+            if (loadedFeedback == null) _dbContext.LearningObjectFeedbacks.Add(feedback);
+            else _dbContext.LearningObjectFeedbacks.Update(feedback);
+
             _dbContext.SaveChanges();
         }
 
-        public void UpdateFeedback(LearningObjectFeedback feedback)
-        {
-            _dbContext.LearningObjectFeedbacks.Update(feedback);
-            _dbContext.SaveChanges();
-        }
-
-        public LearningObjectFeedback GetFeedback(int learningObjectId, int traineeId)
+        public LearningObjectFeedback Get(int learningObjectId, int traineeId)
         {
             return _dbContext.LearningObjectFeedbacks.AsNoTracking().FirstOrDefault(feedback =>
                 feedback.LearningObjectId == learningObjectId && feedback.LearnerId == traineeId);
