+++ SmartTutor/ProgressModel/Feedback/Repository/IFeedbackRepository.cs
@@ -2,8 +2,7 @@ namespace SmartTutor.ProgressModel.Feedback.Repository
 {
     public interface IFeedbackRepository
     {
-        void SaveFeedback(LearningObjectFeedback feedback);
-        void UpdateFeedback(LearningObjectFeedback feedback);
-        LearningObjectFeedback GetFeedback(int learningObjectId, int traineeId);
+        void SaveOrUpdate(LearningObjectFeedback feedback);
+        LearningObjectFeedback Get(int learningObjectId, int traineeId);
     }
 }
