+++ SmartTutor/ProgressModel/ISubmissionService.cs
@@ -1,0 +1,9 @@
+﻿using SmartTutor.ContentModel.LearningObjects.Challenges;
+
+namespace SmartTutor.ProgressModel
+{
+    public interface ISubmissionService
+    {
+        ChallengeEvaluation EvaluateChallenge(string[] sourceCode, int challengeId, string traineeId);
+    }
+}
